+++ Verilog.novaextension/CHANGELOG.md
@@ -1,3 +1,8 @@
+## Version 0.0.2
+
+Added System Verilog to the supported extensions.
+Added basic linting support using verilator.
+
 ## Version 0.0.1
 
 Initial release